Subclass JUnit's BlockJUnit4ClassRunner into ConcurrentRunner
It just overrides the methods of BlockJUnit4ClassRunner necessary to
run test methods annotated with @ConcurrentTest as ordinary test
methods (IOW: methods annotated with @Test).

computeTestMethods() now returns all "ordinary" test methods and all
concurrent test methods in a single list.  To avoid copying and
concatenating the lists of ordinary and concurrent test methods in
each call to computeTestMethods(), we store the resulting concatenated
list in a member variable on the first call.

ConcurrentBuilder now returns a ConcurrentRunner, if the test class
has methods annotated with @ConcurrentTest.

Also add tests to stress computeTestMethods() and verify
ConcurrentBuilder, and put them in a testsuite.

And BTW: these tests can be run by ConcurrentJUnit itself!

Also add a new target to the build script to run the tests.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>