ConcurrentRunner: create ConcurrentFrameworkMethods
computeConcurrentTestMethods() used to simply return a
List<FrameworkMethod> as returned by TestClass.getAnnotatedMethods().
To be able to differenciate "ordinary" test methods (annotated with
@Test) and concurrent test methods (@ConcurrentTest) later, we now
create a ConcurrentFrameworkMethod object for each for each
FrameworkMethod that stands for a concurrent test method, and return
the List of these methods.  Later on the superclass' runChild() method
will get called with these ConcurrentFrameworkMethod objects, and we
will then be able create appropriate concurrent Statements for them.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>