Add ConcurrentStatement class
This class will be responsible to run the concurrent test method in
the environment generating different thread schedules.  But currently
it does nothing of the sort, just behaves like JUnit's InvokeMethod
statement.

Also override BlockJUnit4ClassRunner.methodInvoker() in
ConcurrentRunner class.  This will call the superclass'
methodInvoker(), if the method is an ordinary test method (was
annotated with @Test; FrameworkMethod), or will create a
ConcurrentStatement, if the method is a concurrent test method
(@ConcurrentTest; ConcurrentFrameworkMethod).  Test it, too.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>