Add ExceptionFactory for reconstructing exceptions
In JPF we don't have access to the real exception objects, but we can
find out the exceptions' name, message and stacktrace.
ExceptionFactory is a helper class for reconstructung exceptions based
on that information.

A ceveat: AssertionErrors are somewhat special in the sense that they
don't have a public constructor with a String parameter, but have one
with Object parameter, which will be turned into the detailed message
of the exception.  If an "ordinary" constructor with String parameter
can not be found, we fall back to the one with Object parameter (if
such a constructor exists).

However, exceptions declared in inner classes are not handled and can
not be reconstructed.

Also add tests to check the constructor lookup methods and an ignored
test to illustrate the inner class problem.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>