Test StackFrameConverter.toStackTrace() using JPF
Testing StackFrameConverter.toStackTrace() is difficult, because we
can not just create some of JPF's StackFrame objects (there are
dependencies to other JPF objects).

So, here we add a (rather complex) test, which runs the same small
program with and without JPF, converts the stack trace from JPF, and
compares it with the non-JPF stack trace.

And since this test starts JPF, put it in the new test suite called
JPFTests.  Also modify AllTests test suite to run tests in this new
test suite, too.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>