TestWrapper class with basic functionality
JPF can run only whole applications, IOW it requires a main() method.
The TestWrapper class is used to wrap up individual test methods in a
program, so we can run a test method under JPF.  In the future it will
be extended to handle expected exceptions and invoke @Before methods,
too.

The name of the test class and the test method are supplied on the
command line.  After parsing the command line arguments the test class
is instantiated and the test method is called.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>