TestWrapper: call notifier method on failed test
The test method running under JPF is separated from the unit testing
framework invoking JPF.  Therefore, special means are needed to
communicate the success or failure of a test within JPF with the unit
testing framework.

We will do that by
 * calling the NotifierMethods.testFailed() method from TestWrapper
   when the test fails
 * and monitoring method invocations with a property listener.

If the property listener sees a call to that method, it will report
the failure of the test;  if there is no call to that method, then the
test succeeded.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>