TestWrapper: add expected exception handling
Just as in JUnit, users should be able to test whether a given
exception is thrown by the test method.

This adds the necessary infrastructure (command line argument,
unwrapping the thrown exception from InvocationTargetException, helper
function for checking the exception's type, throwing exception when no
exception was thrown in the test method but one was expected) for
that to the TestWrapper class.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>