TestWrapper: refactor runTest() into runTestMethod()
We are about to add support for @Before methods, and support for
@After methods will follow it shortly, therefore the runTest() method
will have a lot more things to do: run all @Before methods, run the
test method itself and check exceptions, and finally run all @After
methods.  For the sake of simplicity, we will keep all these in
dedicated methods, and runTest() will be just a facade method to
invoke all of them.

To prepare for this, this patch moves the current body of the
runTest() method to the new runTestMethod() method.  Also adjust the
related tests to call runTestMethod() instead of runTest().

Signed-off-by: SZEDER Gábor <szeder@fzi.de>