TestWrapper: fix NPE, consolidate test failure detection
Currently there are three places where we call testFailed(), from
which two is broken.

In the two cases when handling the exception or the lack of an
exception we first call testFailed() and thereafter throw an
exception.  But TestObserver doesn't care about thrown exceptions once
it knows that the test is failed, so these exceptions were
"unnoticed".  This caused the wrong exception to be reported in the
first case and in a NullPointerException in the second.

Here we take a more foolproof approach by moving the testFailed() call
to the caller of runTest(), and call it only if runTest() threw an
exception.  This way testFailed() is always called after the exception
is thrown, so TestObserver will be happy.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>