TestWrapperTest: use custom exception classes
These tests use "ordinary" exception classes like RuntimeException or
AssertionError.  This could lead to confusion between exceptions
thrown by methods of the TestWrapper class and the method tested by
the TestWrapper.  A good example for this is
testExpectingExceptionButNoneIsThrown(), where the test expected an
AssertionError, and the TestWrapper under test also expected an
AssertionError.

To resolve this we will use new custom "testclass-local" exception
classes.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>