Fix exception type when catching different type than expected
If a different exception type is thrown in the test than what is
expected, then JUnit throws an Exception with a specific message
describing the different exception types and using the catched
exception as cause.

Our implementation for concurrent test behave differently, it just
rethrew the catched exception.

With this patch unexpected exceptions in concurrent test methods will
be handled and reported the same way (i.e. throwing Exception, same
message, cause) as in JUnit for ordinary test methods.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>