Handle subclasses of expected exceptions
We checked the type of expected exceptions by simply comparing their
class name with the class name of the actually thrown exception.

This check was too tight, as it only allowed exact type matches,
although a subclass of the expected exception should have been just as
good.  This patch fixes it by using Class.isAssignableFrom() instead.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>