Add OnFailurePublisher
Currently JPF prints if report after each concurrent test, even if the
test actually succeeded, and in those cases the report is not
interesting at all.  It just makes finding the report about a failure
harder, or causes a distracting switch to the console tab in eclipse
if there are no failures.

The OnFailurePublisher class overrides JPF's ConsolePublisher and
modifies its behaviour such a way, that it will only print the report
to the console if the test failed.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>