Show source code in JPF report trace
... to show the developer which source code lines were executed in
which order in which threads when the failure occurred.

This feature is turned on by overriding the appropriate configuration
property through a command line argument.  It would be better to do
this from within the OnFailurePublisher class, but unfortunately the
needed attribute of JPF's ConsolePublisher class is package private.

Note: the source code is only displayed if JPF finds the source files.
However, JPF can not find out where to look for them (neither can
cJUnit), therefore it's necessary to tell it explicitly.  This could
be done by laying a file named 'jpf.properties' (the default filename
where JPF looks for properties) with the contents

  vm.sourcepath = path/to/sources:path/to/other/sources

in the directory where you will run the tests from.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>