ConcurrentDefaultPossibilitiesBuilder: factor out builder list creation
The current ConcurrentDefaultPossibilitiesBuilder.runnerForClass()
method does two things:  creates a list of RunnerBuilders and checks
which one of those is suitable for the testClass.  To add support for
a custom builder this method must be overridden in a subclass, and
that overriding method must again do those two things.

To make it simpler we factor out the creation of the list of builders
into a separate builderList() method, and subclasses will only have to
override it to provide their own boulder list.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>