ConcurrentRunner: refactor Statement builders
Currently all Statement builder method in ConcurrentRunner had an if
statement to distinguish between "ordinary" test methods and
concurrent test methods.  With support for @Before and @After methods
the number of those builder methods, and hence the if statements will
increase, and so will the number of necessary test methods, too.
Therefore, consolidate those if statements in one place, namely in
methodBlock().  This has the additional benefit that there will be one
less levels of indentation in those methods.

While at it, make that one consolidated if statement nicer by using
instanceof instead of Class comparison.

Unfortunately BlockJUnit4ClassRunner.methodBlock() is not that
override-friendly: we had to duplicate significant parts from it in
ConcurrentRunner.methodBlock().  So, put those duplications
into helper methods (createTestObject() and buildStatements()), thus
at least possible subclasses of ConcurrentRunner will be simpler.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>