ExceptionFactoryTest: remove ignored test
Commit 1d5f465f (Add ExceptionFactory for reconstructing exceptions,
2008-12-02) claimed that "exceptions declared in inner classes are not
handled and can not be reconstructed", and added an ignored test case
to illustrate the problem.  Unfortunately, the analysis of the problem
was incorrect.

If the exception is declared in a _static_ inner class, then it can be
reconstructed, moreover, it could be reconstructed even back then.
However, if the exception was declared in a non-static inner class,
then it's entirely impossible to reconstruct it, because we would also
need to reconstruct its outer class.

Therefore, there is no need to keep that ignored test around, so
let's remove it.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>