StackFrameConverter: convert from StackTraceElementInfo[]
In preparation for reconstructing chained exceptions.

This patch adds the methods necessary to convert a
StackTraceElementInfo array into an ordinary StackTraceElement array.

Unfortunately, the file name in stack traces under JPF is different
from the file name in "ordinary" Java stack traces:  it also contains
a path prefix corresponding to the package the file/class belongs to.
Therefore, we still need to strip this prefix with
sourceFileBasename().

Signed-off-by: SZEDER Gábor <szeder@fzi.de>