Add StackTraceElementInfo interface and default implementation
In preparation for reconstructing chained exceptions.

This class will be used to store information about StackTraceElement
instances, both inside and outside of JPF.

A property listener can inspect objects inside JPF relatively easily.
However, to do that it needs to access attributes of objects based on
their names.  But StackTraceElement's attributes are private: their
name are not part of the class' interface, therefore they could change
in future Java versions.  That would break our property listener
badly, as it will not find the attributes it is looking for.

Therefore, we introduce this StackTraceElementInfo interface and its
default implementation, which will just store the relevant attributes
of a StackTraceElement instance in its own attributes, whose name is
known and will not change in the future (or if it will, it will change
together with our property listener).

Signed-off-by: SZEDER Gábor <szeder@fzi.de>