Add ExceptionInfo interface and default implementation
In preparation for reconstructing chained exceptions.

This class will be used to store information about Throwable instances
both inside and outside of JPF; similar to StackTraceElementInfo and
StackTraceElement.  It's called _Exception_Info, because even though
there are actually Throwables, everyone refers them as Exceptions.

For the rationale of why is such a class at all necessary see the log
message of the previous commit adding the StackTraceElementInfo
interface and its default implementation class.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>