TestObserver: change the way exceptions are observed
The current exception observing mechanism listens in the
TestObserver.exceptionThrown() method, which will be each time an
exception is thrown.  It collects information about all exception
thrown during the test run until the test fails;  a newer exception
overwrites the info stored about the previous one, and TestObserver
will report the last detected exception at the end of the test.
Unfortunately, this way we don't know anything about the cause of the
exception (as in chained exceptions), hence we can't reconstruct an
exception with a cause properly (iow.: we can't reconstruct its
cause).

To resolve this, this patch starts using all the stuff added in the
previous commit in order to store the information about the exception
thrown in the test in an ExceptionInfo instance.  Since ExceptionInfo
also holds information about the cause of the exception and
ExceptionFactory can handle that, we will be able to reconstruct
exception chains, too.

If an exception is thrown during reconstructing, it will be stored and
later returned by getException(), so it will be reported to the unit
testing framework properly (otherwise it would be caught in JPF.run(),
its stack trace would be printed, and the unit testing framework will
never know about it).

Signed-off-by: SZEDER Gábor <szeder@fzi.de>