TestObserverTest: test exception thrown during collecting exception info
An exception might be thrown in a property listener while TestObserver
collects information about the exception thrown by the observed test.
Such an exception is catched by JPF and is not rethrown, so it will
not reach the callers of JPF.  As a workaround we report this
exception through the same means used for reporting exceptions thrown
by the test.

This patch adds a test to check that this actually works as expected.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>