Add TestProperty interface
This provides a unified interface for all search properties observing
unit tests run under JPF.  It declares methods to query the result of
the test and the exception that should be reported to the user in case
of a failed test.

Developers, who want to verify their unit tests with custom properties
and need to report special exceptions in case of a failed test will
have to implement this interface in their properties.  We will also do
so with our TestObserver in the following patches.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>