JPFInvoker: make some members and methods protected
The methods runJPF() and createJPFArgs() were public, but they are
more like helper methods and should not be part of the class' public
interface.  Furthermore, there are other helper methods and member
variables that are package-private.

This patch makes all these methods and member variables protected, to
be nice to those who might ever want to subclass this class.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>