OnFailurePublisher: query for test failure at JPFInvoker
OnFailurePublisher currently queries only the TestObserver property to
find out whether the test failed or not.  This patch changes it to
query JPFInvoker instead using the getTestResult() method added in the
previous commit.  This way it will be able to act properly if an other
property is violated during the test run.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>