ElementInfoWrapper: compatibility with JPF's SVN trunk
So far we used the 'v5' branch from the JPF Subversion repository.
However, the main development is done on trunk, and it contains some
improvements/bugfixes that are important for us (e.g. v5 apparently
does not explore all possible states in Test22 from the Helgrind test
suite, while trunk does).  Therefore, we need to switch to use JPF's
trunk at r1539.

Unfortunately, there are a few commits on v5 that are not on trunk,
and out ElementInfoWrapper class relies on the two methods added by
those commits.  This patch eliminates the dependency on those commits,
so cJUnit will work with JPF's trunk and v5 branch as well.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>