TestObserver.getException(): handle exception during
TestObserver.getException() might throw any kind of exceptions related
to instantiation using reflection.  However, it will soon be converted
to implement the TestProperty interface, which defines a
getException() method that does not throw anything.

To resolve this, getException() will catch and return the exception
that is thrown during exception instantiation, and this returned
exception will then be reported to user instead of the exception that
actually caused the test to fail.  Of course, the user is interested
in the exception that caused the test to fail, but if the
instantiation of that exception outside of JPF failed with an
exception, then we have no chance to report the failure-causing
exception anyway...

Signed-off-by: SZEDER Gábor <szeder@fzi.de>