JPFInvoker: checkProperties() retrieves the violated property from JPF
Currently checkProperties() only queries whether TestFailedProperty
was violated, but not any other properties.  With this patch it will
get the violated property from JPF and report an exception according
to the violated property using the recently added
getExceptionFromProperty() method.

This makes possible to report the violation of properties other than
TestFailedProperty, e.g. to report the violation of
NotDeadlockedProperty in case of a deadlock.

Also modify some related tests to instantiate a JPFInvoker that
delivers proper Error list for testing.  Furthermore, remove the now
unnecessary getTestResult() overrides.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>