JPFInvoker: handle multiple property violations
Although propably not a common case, but it is possible that multiple
properties got violated at the same time.

This patch handles this case by throwing a MultipleFailureException
containing the exceptions collected from all the violated properties.

The test added for this feature only works with Hamcrest 1.1 at the
moment, and even that requires some ugly casting magic to compile with
a non-eclipse-supplied Java compiler.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>