TestObserver: handle null callee
In TestObserver.handleInvokeInstruction() we assumed that
InvokeInstruction.getInvokedmethod() always returns a valid object as
callee.  However, it can also return a null reference if the observed
program performs arithmetic with a variable whose type is a subclass
of java.lang.Number and which is a null reference.  We didn't check
the return value for null and were "rewarded" with a
NullPointerException.

This patch adds the necessary check.  Since we are only interested in
calls to methods of the NotifierMethods class, and for those calls
getInvokedMethod() always returns a valid object, we just ignore those
cases when we got a null reference.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>