ConcurrentBuilder: return ConcurrentRunner from runnerForClass()
ConcurrentBuilder.runnerForClass() returns either a null reference or
an instance of the ConcurrentRunner class, which is a subclass of
Runner.  We could as well show this in the method signature, so change
the return type from Runner to ConcurrentRunner.

Also change the type in the related tests, so we don't have to assert
the returned type of the returned object explicitly, because it will
be checked by the compiler.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>