TestObserverTest: use instanceOf() instead of is() when checking types
Hamcrest's is() method provides a shortcut for the frequently used
is(instanceOf(SomeClass.class)) call.  We use this shortcut in two
tests for our TestObserver class.

Unfortunately, Hamcrest v1.2 has tightened the definition of is() for
this shortcut in an incompatible way, such that the argument of is()
must be a superclass of the variable to be asserted.  Since we test
the other way around (we test that an instance of a superclass is
indeed an instance of one of its subclasses), the test does not
compile anymore.

This patch turns those two is() calls to instanceOf(), so the tests
will compile again.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>