TestWrapper: fix visibility
All of TestWrapper's methods were declared as public, although they
were not meant to be called from outside of this class.  The
attributes of this class, however, were declared as package-private.

So, let's change their visibility to protected, to be nice to people
who eventually want to subclass this class in the future.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>