TestWrapper: helper method to invoke methods through reflection
To check the exception thrown during the test method, runTestMethod()
needs an InvocationTargetException with the fault-causing exception as
its cause.  However, exceptions thrown by @Before and @After methods
don't need to be checked.

To make the invocation of @Before and @After methods easier, this
patch adds the helper method invokeMethodUnchainingException(), which
will unwrap the fault-causing exception from the
InvocationTargetException.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>