TestFailedProperty: don't register as property when search starts
JPF's PropertyListenerAdapter.searchStarted() automatically registers
the instance as property at the Search instance.  Our
TestFailedProperty class extends this class, and leaves this behaviour
intact.  As a result, TestFailedProperty is registered twice as
Property, because JPFInvoker already registered it as PropertyListener
before the search started.

To avoid this double registration this patch overrides that
searchStarted() method to make it a no-op.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>