Revert "JPFInvoker: handle multiple property violations"
This reverts commit f744d0a217b1596dd469a9c6e3e35ca3cb76c4a1.

As shown by the test
AssumptionsAboutJPF.multiplePropertiesViolatedAtTheSameTime(), if
multiple properties are violated at the same time, JPF only reports
the first of them.  Therefore, the code added in the reverted commit
will never trigger (well, except when searching for multiple errors,
but that is a completely different story, and the rest of our system
can't handle it properly anyway).

Signed-off-by: SZEDER Gábor <szeder@fzi.de>