Counter helper class for testing
Sometimes we want to check whether a method was invoked or not during
a test without the additional dependency of some fancy mocking
framework.  To do that we usually need something like the following:

	public void testMethod() {
        	int invoked;
		ClassToTest c = new ClassToTest() {
			@Override public void someMethod() {
				invoked++;
			}
		};
		// test, test, test
		assertThat(invoked, equalTo(1));
	}

Unfortunately, this does not work, because a method of an anonymous
inner class defined in a method of the outer class can't modify a
variable declared in the method of the outer class, unless that
variable is declared as final.  But if it's declared as final, then
new values can't be assigned to it.  Moving the variable out of the
method to turn it into a member variable is very ugly, especially if
we have multiple such tests in the same class.

So, to make things slightly cleaner, this patch introduces a trivial
Counter class, which is basically a wrapper around an int, and can be
increment()ed even if it's declared as final.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>