Add integration tests to the test suite
... to see that the units are working as expected when they are wired
together, too.  These tests we invoke ConcurrentJUnit.runClasses(),
which is cJUnit's main entry point when it's used as a library.

There are in total seven test cases that we run in these integration
tests:

 * a succeeding sequential test case
 * a succeeding concurrent test case
 * a succeeding concurrent test case with expected exception
 * a failing sequential test case
 * a concurrent test case with a sequential bug
 * a concurrent test case with a datarace
 * a concurrent test case with a deadlock

The first three from these are checked from the same integration test,
so we also check that sequential and concurrent test cases can be
mixed in the same test class.  The latter four are checked by
different integration tests, because if they would be in the same test
class, then the order of their invocation is unspecified, and we
wouldn't know which failure comes from which test case in the results.

System.out is disabled (by redirecting it to a ByteArrayOutputStream
which will then be thrown away) for these tests, because JPF pollutes
stdout by reporting the error in the buggy concurrent test cases.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>