TestFailedProperty: reconstruct exception already in testFailed()
So we don't store the exceptionInfo anymore, but the reconstructed
exception.  It makes the code more consistent, because when collecting
exception info fails with an exception, then we store this exception
itself and not some exceptioninfo.  Furthermore, we will need to
reconstruct the exception anyway during reporting, so why not
reconstruct it right away.  It also greatly simplifies the
getException() method, as we don't need that lazy initialization like
construct in there.  Last, we implement it as a separate helper
method, making testing of testFailed() easier.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>