Notify TestFailedProperty about a succeeded test run
Currently TestFailedProperty is only interested in the failure of the
test, what it gathers by observing the call of the
NotifierMethods.testFailed() method.  To distinguish between
concurrency and ordinary bugs it will also need to know when a test
have succeeded in an execution path.

This patch adds the new notifier method testSucceeded() that will be
called from TestWrapper when the test succeeded, and extends
TestFailedProperty.handleMethodInvocation() to observe that call and
invoke testSucceeded().

Signed-off-by: SZEDER Gábor <szeder@fzi.de>