TestFailedProperty: found failed/succeeded path variables and methods
If we find an execution path in which the test succeeds before any of
the properties are violated, then we are dealing with a concurrency
bug.  To identify this case, ResultCollector needs to know whether
TestFailedProperty already observed at least one success before the
property violation occurred.

After this patch TestFailedProperty can be queried whether it saw at
least one successful test run.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>