ResultCollector: report deadlocks as DeadlockError
This is done by checking whether the first violated property is
NotDeadlockedProperty.

Deadlocks are concurrency bugs, even if they manifest the exact same
way in all execution paths, because they involve synchronization
primitives.  Therefore, if ResultCollector sees that the first
violated property is NotDeadlockedProperty, then it immediately stops
the search and reports it as a failed test, without checking for
different results in other execution paths.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>