ResultCollector: identify concurrency bugs
We will search for multiple errors until we can determine whether the
test fails because of a concurrency bug, fails because of a sequential
bug, or succeeds in all execution paths.

The test fails because of a concurrency bug, if we can find two execution
paths which lead to different results, i.e. the test fails in on
execution path while succeeds in an other, or it fails differently in
different execution paths.  A test fails differently in two execution
paths if either different properties are violated or different
exceptions are thrown.

The test fails because of a sequential bug, if it fails the same way in all
execution paths, i.e. same violated property, same exception.

To keep reporting simple for now, only the first noticed failure is
reported if we find execution paths leading to different failures.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>