TestFailedProperty: add means to report success as failure
If a test fails in the first examined execution path and succeeds in
an other one, then there is a concurrency bug and we can stop the
search.  However, ResultCollector only sees property violations, so it
can't figure out when a test succeeds.

Therefore, this patch adds the ability to TestFailedProperty to report
a the success of a test as failure, thereby making the success
noticeable for ResultCollector.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>