Add ExceptionComparator class
Throwable's and its subclasses' equals() method only compares the
object references, hence it could only tell whether the two objects
are actually the same.  This is not enough for us to compare
exceptions thrown in different execution paths, therefore this patch
adds a class to compare two Throwable instances thoroughly.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>