Matcher and helper class for testing critical sections
'wait()' and 'notify()/notifyAll()' are often used to control
exclusive access to resources based on some condition.  Testing
whether the condition's evaluation and the notification of waiting
threads in a concurrent unit test is rather difficult, because the
developer has to check that not more than one thread executes the code
between the condition and the notification.

With the 'accessedExclusively()' matcher introduced in this patch this
will boil down to an assertion, as shown in the example below.

  @ConcurrentTest
  public void testExclusiveAccess() throws Throwable {
      final AccessManager am = new AccessManager();
      final CriticalSection cs = new CriticalSection();
      Thread t = new Thread() {
          public void run() {
              am.waitUntilAllowed();
              assertThat(cs, accessedExclusively());
              am.notifyWaiters();
          }
      };
      t.start();
      am.waitUntilAllowed();
      assertThat(cs, accessedExclusively());
      am.notifyWaiters();
      t.join();
  }

Signed-off-by: SZEDER Gábor <szeder@fzi.de>