TestWrapper: method and exception as suboptions
The API for concurrent unit tests will allow developers to mark
methods that need to be executed simultaneously in separate threads.
This will require to specify multiple test methods and expected
exceptions on TestWrapper's command line.  However, currently we have
different options for test methods and expected exceptions, which
makes pairing them up difficult.

This patch changes TestWrapper's command line options in a way that
makes it obvious which expected exception belongs to which test
method: the test method and expected exception will be specified as
suboptions of the test options, e.g.:

  --test=method=testMethod,exception=ExpectedException

Signed-off-by: SZEDER Gábor <szeder@fzi.de>