TestWrapperTest: call runTestMethod() instead of invokeTestMethod()
invokeTestMethod() is just a wrapper for invoking the test method
through reflection, and is an implementation detail of
runTestMethod().  To test that the test method is not invoked in
testMethodIsNotInvokedIfABeforeMethodFails() it is enough to check
that the runTestMethod() method is not invoked, and is not necessary
for the test to check as deep as invokeTestMethod().

Signed-off-by: SZEDER Gábor <szeder@fzi.de>