TestWrapper: make checkExceptions() list-aware
... so it will check the expected exception of all TestMethods.

Since multiple TestMethods can throw an exception or have expected
exception specifications, all of them must be verified.  With this
patch checkExceptions() will do that, and any unexpected exceptions
will be appended to the list of errors.

This change has no visible effect in case of a basic concurrent unit
test taking care of starting-stopping the test's threads in itself.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>