TestWrapper: invoke multiple TestMethods in separate threads
This is the last change to prepare TestWrapper for APIs for concurrent
unit tests that allow developers to mark methods that need to be
executed simultaneously in separate threads.

Care is taken not to use unnecessary threads (to ease the work of
JPF), i.e. the first TestMethod is invoked in the main thread, while
dedicated threads are created for every other TestMethods.

Like the previous patches in this series, this change has no visible
effect in case of a basic concurrent unit test taking care of
starting-stopping the test's threads in itself.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>