TestFailedProperty: split handleMethodInvocation()
For some unknown reason, creating a MethodInfo instance results in a
NPE, if JPF was not initialized previously.  There are a few test
cases testing TestFailedProperty.handleMethodInvocation() (namely
handleMethodInvocationWhenNotNotifierMethod(),
handleMethodInvocationInvokesTestFailed(), and
handleMethodInvocationInvokesTestSucceededOnSuccess() in
TestFailedProeprtyTest) that are affected by this issue, and fail if
the test class is run on its own (i.e. not as part of the full cJUnit
test suite), or if these test cases are run by the continuous testing
plugin.

Since the issue is rooted in JPF, we can only workaround it by
removing as much dependency on MethodInfo instances from test cases as
possible.

The first step of this workaround is to make most of the login in
handleMethodInvocation() testable without a MethodInfo instance.
Therefore, this patch splits out that login into a dedicated method,
which gets the callee's class name and method name as parameters,
instead of the MethodInfo instance representing the callee.  This will
allow us to modify the related tests to use just Strings for input
instead of MethodInfo instances.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>