JPFInvokerTest: remove getJPFSearchErrorsReturnsUnalteredErrorList() test
This test case checks whether JPFInvoker.getJPFSearchErrors() returns
the same error list on subsequent invocations.  However,
JPFInvoker.getJPFSearchErrors() is just a one line long method
wrapping JPF.getSearchErrors(), so there is not much that can go
wrong.  Despite the simplicity of the method under test, the test case
is rather complex and slow, because it involves the initialization of
JPF, causing subtle issues when we run JPF later with different
configurations.

All in all, we are better off without this test case.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>