Delete JPFTests test suite
As it turned out, running JPF subsequently with different
configurations in the same JVM is prone to subtle errors.  Since the
tests checking our assumptions about JPF will indeed run JPF with
different configurations, we can't rely on the results of the tests if
they are run as part of the JPFTests test suite.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>