Move arg creation from JPFInvoker into ConcurrentStatement
Currently ConcurrentStatement passes a bunch of parameters to
JPFInvoker, which in turn is reponsible for creating the "command
line" arguments for running TestWrapper inside JPF.

By extending the API for concurrent unit tests the number of
parameters needed to create the command line arguments will increase.
Instead of passing the growing number of parameters from
ConcurrentStatement to JPFInvoker, let's move argument creation to
ConcurrentStatement, where all the information needed for the
arguments will be gathered anyway.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>