ConcurrentFrameworkMethod: add method to easily get expected exception
Currently, ConcurrentRunner queries the FrameworkMethod instance about
the annotations the method is marked with, whether there is
expected exception specification among those annotations, and what is
that expected exception.

But conceptually this should belong to the FrameworkMethod class.
However, since FrameworkMethod is part of JUnit, we can't change it,
so let's add a method to query the expected exception at to our
ConcurrentFrameworkMethod class at least.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>