ConcurrentRunner: handle the threadCount annotation parameter
Finally, after all the preparations done in the previous commits, this
patch "switches on" the threadCount feature, which allows developers
to specify the number of threads that will concurrently invoke the
marked method.

Its implementation is really straigtforward: it adds the
ConcurrentFrameworkMethod to the ConcurrentStatement as many times as
specified by the threadCount annotation parameter.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>